refactor: ensure that reducer is pure function
