--- src/data/userState.js
@@ -1,10 +1,0 @@
-export const initialUserState = {
-  users: {
-    admin: 'iamtheboss',
-    olhanotolga: 'bestpassword',
-  },
-  userData: {
-    username: '',
-    password: '',
-  },
-};
