+++ src/reducers/userReducer.js
@@ -1,5 +1,3 @@
-import { initialUserState } from '../data/userState';
-
 export const ACTION_TYPES = {
   LOGGED_IN: 'logged-in',
   SIGNED_UP: 'signed-up',
@@ -8,6 +6,27 @@ export const ACTION_TYPES = {
 
 export function userReducer(state, action) {
   const { type, payload } = action;
+
+  // Because reducers are pure functions, I cannot use any external values
+  // Should be same as in UserContext
+  const initialUserState = {
+    users: {
+      admin: 'iamtheboss',
+      olhanotolga: 'bestpassword',
+    },
+    userData: {
+      username: '',
+      password: '',
+    },
+  };
+
+  // Same here, I should see that this const is same as exported pseudo-enum at the top
+  const ACTION_TYPES = {
+    LOGGED_IN: 'logged-in',
+    SIGNED_UP: 'signed-up',
+    LOGGED_OUT: 'logged-out',
+  };
+
   switch (type) {
     case ACTION_TYPES.LOGGED_IN: {
       return {
